+++ Client/Changelog.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 1.0.9
+
+### Новое
+* Добавлена страница audio-editor.aspx
+
+### Исправления
+* Обновление версий продуктов avs
+* Заменены изображаения в слайдах ` Free upgrades / Unlimited access to the program `, на страницах: installed-audio-converter.aspx, installed-audio-editor.aspx, installed-pack.aspx, installed-video-converter.aspx, installed-video-editor.aspx, installed-video-remaker.aspx
+* Заменены html теги для заголовков, на страницах avs-free-video-converter.aspx/avs-free-audio-converter.aspx для языков fr/it: `Convert and edit all key audio formats` H2 на H3, `Technical specification` - p
+
 ## 1.0.8
 
 ### Исправления